SlideGod/requirements.txt
python-pptx>=0.6.21
requests>=2.31.0
openai>=1.0.0
typing-extensions>=4.5.0
pillow>=9.5.0  # Required for image handling in python-pptx
lxml>=4.9.0    # Required for XML processing in python-pptx 